requirements.txt
opensearch-py>=2.4.0
requests>=2.31.0
apscheduler>=3.10.4
pyyaml>=6.0.1
numpy>=1.26.0
python-dotenv>=1.0.0
geoip2>=4.8.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
rich>=13.7.0
click>=8.1.7
fastapi>=0.115.0
uvicorn>=0.32.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
